% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{as.data.frame.csvw_table}
\alias{as.data.frame.csvw_table}
\title{Convert csvw_table to data.frame S3 method}
\usage{
\method{as.data.frame}{csvw_table}(x, ...)
}
\arguments{
\item{x}{csvw_table object}

\item{...}{extra args}
}
\value{
data.frame
}
\description{
Convert csvw_table to data.frame S3 method
}
