% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{as.data.frame.csvw_table_group}
\alias{as.data.frame.csvw_table_group}
\title{Convert csvw_table_group to a list of data.frames}
\usage{
\method{as.data.frame}{csvw_table_group}(x, ...)
}
\arguments{
\item{x}{csvw_table_group object}

\item{...}{extra args}
}
\value{
list of data.frames
}
\description{
Convert csvw_table_group to a list of data.frames
}
