% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{check_primary_key}
\alias{check_primary_key}
\title{Check uniqueness of primary key columns in table rows}
\usage{
check_primary_key(table, rows)
}
\arguments{
\item{table}{csvw_table object}

\item{rows}{list of row structures}
}
\value{
logical
}
\description{
Check uniqueness of primary key columns in table rows
}
