% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{check_referential_integrity}
\alias{check_referential_integrity}
\title{Validate foreign key constraints in table group}
\usage{
check_referential_integrity(t_obj, strict = FALSE)
}
\arguments{
\item{t_obj}{csvw_table_group object}

\item{strict}{logical, if TRUE raises warning/error for nulls in composite foreign keys}
}
\value{
logical
}
\description{
Validate foreign key constraints in table group
}
