% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_path}
\alias{clean_path}
\title{Helper to clean path relative movements like /./ and /../}
\usage{
clean_path(path)
}
\arguments{
\item{path}{path string}
}
\value{
cleaned path
}
\description{
Helper to clean path relative movements like /./ and /../
}
