% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{csvw}
\alias{csvw}
\title{CSVW constructor}
\usage{
csvw(url, md_url = NULL, validate = FALSE, lax = FALSE)
}
\arguments{
\item{url}{path or URL of CSV file}

\item{md_url}{optional explicit metadata path or URL}

\item{validate}{logical, if TRUE validates constraints and referential integrity}

\item{lax}{logical, if TRUE ignores unknown metadata properties rather than flagging them}
}
\value{
a csvw S3 object
}
\description{
CSVW constructor
}
