% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dialect.R
\name{csvw_dialect}
\alias{csvw_dialect}
\title{Create a CSVW Dialect object}
\usage{
csvw_dialect(
  commentPrefix = NULL,
  delimiter = ",",
  doubleQuote = TRUE,
  encoding = "utf-8",
  header = TRUE,
  headerRowCount = 1,
  lineTerminators = NULL,
  quoteChar = "\\"",
  skipBlankRows = FALSE,
  skipColumns = 0,
  skipInitialSpace = FALSE,
  skipRows = 0,
  trim = FALSE,
  ...
)
}
\arguments{
\item{commentPrefix}{character prefix for comments}

\item{delimiter}{character field separator}

\item{doubleQuote}{logical indicating whether quote chars are doubled}

\item{encoding}{character encoding}

\item{header}{logical indicating if file has header row}

\item{headerRowCount}{integer number of header rows}

\item{lineTerminators}{character vector of line terminators}

\item{quoteChar}{character used for quoting}

\item{skipBlankRows}{logical indicating if blank rows are skipped}

\item{skipColumns}{integer number of columns to skip at start of row}

\item{skipInitialSpace}{logical indicating if initial spaces are skipped}

\item{skipRows}{integer number of rows to skip at start of file}

\item{trim}{logical or character specifying how values are trimmed}

\item{...}{other attributes}
}
\value{
a csvw_dialect object
}
\description{
Create a CSVW Dialect object
}
