% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expand_uri_template}
\alias{expand_uri_template}
\title{Expand RFC 6570 URI template using context variables}
\usage{
expand_uri_template(template, context)
}
\arguments{
\item{template}{URI template string}

\item{context}{list containing variables for expansion}
}
\value{
expanded string
}
\description{
Expand RFC 6570 URI template using context variables
}
