% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_json}
\alias{get_json}
\title{Fetch and parse a JSON document from local path or URL}
\usage{
get_json(url_or_path)
}
\arguments{
\item{url_or_path}{path or URL}
}
\value{
list structure
}
\description{
Fetch and parse a JSON document from local path or URL
}
