% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_table}
\alias{get_table}
\title{Access a table by the last component of its URL}
\usage{
get_table(x, name, ...)
}
\arguments{
\item{x}{a csvw object}

\item{name}{character string of the last component of the table's URL (e.g. filename)}

\item{...}{extra args}
}
\value{
data.frame containing the table rows, or NULL if not found
}
\description{
Access a table by the last component of its URL
}
