% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{inherit_null}
\alias{inherit_null}
\title{Helper to inherit null values}
\usage{
inherit_null(col)
}
\arguments{
\item{col}{column object}
}
\value{
character vector of null representations
}
\description{
Helper to inherit null values
}
