% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{inherit_val}
\alias{inherit_val}
\title{Helper to inherit value from parent structure}
\usage{
inherit_val(col, name)
}
\arguments{
\item{col}{column object}

\item{name}{property name}
}
\value{
value or default
}
\description{
Helper to inherit value from parent structure
}
