% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{locate_metadata}
\alias{locate_metadata}
\title{Locate metadata matching CSVW specifications}
\usage{
locate_metadata(url, md_url = NULL)
}
\arguments{
\item{url}{path or URL of CSV file}

\item{md_url}{optional explicit metadata path or URL}
}
\value{
list with parsed JSON md structure and no_header logical
}
\description{
Locate metadata matching CSVW specifications
}
