% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{parse_column}
\alias{parse_column}
\title{Parse Column list structure}
\usage{
parse_column(c, parent = NULL, generated = FALSE)
}
\arguments{
\item{c}{list representing column}

\item{parent}{parent Table object}

\item{generated}{logical, if TRUE represents an autogenerated column}
}
\value{
a csvw_column object
}
\description{
Parse Column list structure
}
