% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatypes.R
\name{parse_datatype}
\alias{parse_datatype}
\title{Parse list structure or string to a Datatype definition}
\usage{
parse_datatype(d)
}
\arguments{
\item{d}{character string or list representing datatype}
}
\value{
datatype list structure
}
\description{
Parse list structure or string to a Datatype definition
}
