% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatypes.R
\name{parse_datatype_value}
\alias{parse_datatype_value}
\title{Helper to parse base datatype value}
\usage{
parse_datatype_value(v, base_type, format = NULL)
}
\arguments{
\item{v}{character string}

\item{base_type}{base datatype name}

\item{format}{optional format specification}
}
\value{
parsed R value
}
\description{
Helper to parse base datatype value
}
