% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatypes.R
\name{parse_datetime}
\alias{parse_datetime}
\title{Parse datetime formats}
\usage{
parse_datetime(v, format, base_type)
}
\arguments{
\item{v}{raw string}

\item{format}{format pattern}

\item{base_type}{date or datetime base type}
}
\value{
Date or POSIXct
}
\description{
Parse datetime formats
}
