% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatypes.R
\name{parse_decimal}
\alias{parse_decimal}
\title{Parse decimal values from string supporting group and decimal characters}
\usage{
parse_decimal(v, format, base_type = "decimal")
}
\arguments{
\item{v}{raw string}

\item{format}{format list/string}

\item{base_type}{base datatype name}
}
\value{
numeric
}
\description{
Parse decimal values from string supporting group and decimal characters
}
