% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dialect.R
\name{parse_dialect}
\alias{parse_dialect}
\title{Parse Dialect list structure}
\usage{
parse_dialect(d, fallback = NULL, base_path = "")
}
\arguments{
\item{d}{list or csvw_dialect object}

\item{fallback}{fallback dialect object}

\item{base_path}{base path/URL for resolving relative paths}
}
\value{
a csvw_dialect object
}
\description{
Parse Dialect list structure
}
