% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{parse_table}
\alias{parse_table}
\title{Parse Table list structure}
\usage{
parse_table(t, parent = NULL, base_path = "")
}
\arguments{
\item{t}{list representing table}

\item{parent}{parent TableGroup object}

\item{base_path}{character directory path or base URL}
}
\value{
a csvw_table object
}
\description{
Parse Table list structure
}
