% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{print.csvw}
\alias{print.csvw}
\title{S3 Print method for csvw}
\usage{
\method{print}{csvw}(x, ...)
}
\arguments{
\item{x}{csvw object}

\item{...}{extra print args}
}
\description{
S3 Print method for csvw
}
