% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{read_csvw}
\alias{read_csvw}
\title{Convenient helper to read CSVW tables}
\usage{
read_csvw(url, md_url = NULL)
}
\arguments{
\item{url}{path or URL}

\item{md_url}{optional metadata URL or path}
}
\value{
list of data.frames (or single data.frame if only one table exists)
}
\description{
Convenient helper to read CSVW tables
}
