% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{read_table_csv}
\alias{read_table_csv}
\title{Read data for a CSVW table}
\usage{
read_table_csv(table, strict = TRUE, validate = FALSE, lax = FALSE)
}
\arguments{
\item{table}{csvw_table object}

\item{strict}{logical, if TRUE throws on validation errors}

\item{validate}{logical, if TRUE performs validation checks during reading}

\item{lax}{logical, if TRUE ignores excess columns / compatibility issues during lax validation}
}
\value{
list of row lists
}
\description{
Read data for a CSVW table
}
