% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resolve_url}
\alias{resolve_url}
\title{Resolve path or URL relative to base URL or path}
\usage{
resolve_url(base, ref)
}
\arguments{
\item{base}{character string base URL or directory path}

\item{ref}{character string relative path or URL}
}
\value{
resolved path or URL
}
\description{
Resolve path or URL relative to base URL or path
}
