% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonld.R
\name{to_json.csvw}
\alias{to_json.csvw}
\title{Convert csvw S3 object to JSON structure}
\usage{
\method{to_json}{csvw}(x, minimal = FALSE, ...)
}
\arguments{
\item{x}{csvw object}

\item{minimal}{logical}

\item{...}{extra args}
}
\value{
list structure
}
\description{
Convert csvw S3 object to JSON structure
}
