% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonld.R
\name{to_json}
\alias{to_json}
\title{Convert CSVW object to JSON representation}
\usage{
to_json(x, minimal = FALSE, ...)
}
\arguments{
\item{x}{csvw object}

\item{minimal}{logical, if TRUE returns flat list of descriptors, else standard group/table structure}

\item{...}{extra args}
}
\value{
list structure ready for JSON serialization
}
\description{
Convert CSVW object to JSON representation
}
