% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatypes.R
\name{validate_constraints}
\alias{validate_constraints}
\title{Validate datatype constraints}
\usage{
validate_constraints(val, dt)
}
\arguments{
\item{val}{parsed R value}

\item{dt}{datatype definition list}
}
\value{
logical TRUE if constraints met, errors otherwise
}
\description{
Validate datatype constraints
}
