% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{validate_csvw}
\alias{validate_csvw}
\title{Validator function for CSVW}
\usage{
validate_csvw(url, md_url = NULL, lax = FALSE)
}
\arguments{
\item{url}{path or URL}

\item{md_url}{optional metadata path or URL}

\item{lax}{logical, if TRUE ignores metadata properties that do not strictly comply}
}
\value{
logical TRUE if valid
}
\description{
Validator function for CSVW
}
