% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{write_csvw}
\alias{write_csvw}
\title{Write data for all tables in a TableGroup}
\usage{
write_csvw(tg, fname, strict = FALSE, ...)
}
\arguments{
\item{tg}{csvw_table_group or csvw_table}

\item{fname}{filename for the metadata file}

\item{strict}{logical, if TRUE raises error if data contains fields not in schema}

\item{...}{other args}
}
\value{
number of rows written or NULL
}
\description{
Write data for all tables in a TableGroup
}
