% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{does_metadata_describe_file}
\alias{does_metadata_describe_file}
\title{Check if metadata describes the CSV file}
\usage{
does_metadata_describe_file(md, url)
}
\arguments{
\item{md}{list representing parsed metadata}

\item{url}{path or URL of CSV file}
}
\value{
logical
}
\description{
Check if metadata describes the CSV file
}
